Use std::optional for random_partial return value
There is a difference between no partial path and an empty store.
An empty store can occur when the execution point leading to the
path does not depend on symbolic values. In this case we still
want to re-execute and hence it is not sufficient to return
an empty store.